refactor(edit-user.component.ts): Moved edit user to the user directory and changed it into a standalone component
