+++ src/BurstChat.Client/src/app/components/user/edit-user/edit-user.component.html
@@ -1,25 +1,13 @@
 <div class="burst-container-fluid bg-secondary">
-
     <burst-card>
-
         <burst-card-body>
-
-            <burst-edit-user-info></burst-edit-user-info>
-
+            <burst-edit-user-info/>
         </burst-card-body>
-
         <burst-card-body>
-
-            <burst-edit-user-invitations></burst-edit-user-invitations>
-
+            <burst-edit-user-invitations/>
         </burst-card-body>
-
         <burst-card-body>
-
-            <burst-edit-user-media></burst-edit-user-media>
-
+            <burst-edit-user-media/>
         </burst-card-body>
-
     </burst-card>
-
 </div>
