+++ src/BurstChat.Client/src/app/components/user/edit-user/edit-user.component.scss
