+++ src/BurstChat.Client/src/app/modules/user/user.module.ts
@@ -6,7 +6,6 @@ import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { library } from '@fortawesome/fontawesome-svg-core';
 import { faUserCircle } from '@fortawesome/free-solid-svg-icons';
 import { UserRoutingModule } from 'src/app/modules/user/user.routing';
-import { EditUserComponent } from 'src/app/modules/user/components/edit-user/edit-user.component';
 import { CardComponent } from 'src/app/components/shared/card/card.component';
 import { CardHeaderComponent } from 'src/app/components/shared/card-header/card-header.component';
 import { CardBodyComponent } from 'src/app/components/shared/card-body/card-body.component';
@@ -18,6 +17,7 @@ import { ExpanderComponent } from 'src/app/components/shared/expander/expander.c
 import { EditUserMediaComponent } from 'src/app/components/user/edit-user-media/edit-user-media.component';
 import { EditUserInvitationsComponent } from 'src/app/components/user/edit-user-invitations/edit-user-invitations.component';
 import { EditUserInfoComponent } from 'src/app/components/user/edit-user-info/edit-user-info.component';
+import { EditUserComponent } from 'src/app/components/user/edit-user/edit-user.component';
 
 @NgModule({
     imports: [
@@ -36,12 +36,9 @@ import { EditUserInfoComponent } from 'src/app/components/user/edit-user-info/ed
         ExpanderComponent,
         EditUserMediaComponent,
         EditUserInvitationsComponent,
-        EditUserInfoComponent
-    ],
-    declarations: [
-        EditUserComponent,
-    ],
-    providers: []
+        EditUserInfoComponent,
+        EditUserComponent
+    ]
 })
 export class UserModule {
