+++ src/BurstChat.Client/src/app/modules/user/user.routing.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { EditUserComponent } from 'src/app/modules/user/components/edit-user/edit-user.component';
+import { EditUserComponent } from 'src/app/components/user/edit-user/edit-user.component';
 
 const routes: Routes = [
     {
